Added Threads logics to upload file

+++ src/main/java/com/fran/pruebams/application/service/FileService.java
@@ -14,8 +14,16 @@ import java.nio.file.Paths;
 
 @Service
 public class FileService {
+
+    public boolean fileExists(MultipartFile file) {
+        String directory = "files";
+        Path dirPath = Paths.get(directory);
+        Path path = Paths.get("files/" + file.getOriginalFilename());
+        return Files.exists(path);
+    }
+
     @Async
-    public void saveFile(MultipartFile file) {
+    public void saveFile(MultipartFile file) throws com.fran.pruebams.application.exceptions.FileAlreadyExistsException {
         try {
             PipedInputStream in = new PipedInputStream();
             PipedOutputStream out = new PipedOutputStream(in);
@@ -29,6 +37,9 @@ public class FileService {
             Thread writerThread = new Thread(() -> {
                 try (in) {
                     Path path = Paths.get("files/" + file.getOriginalFilename());
+                    if (Files.exists(path)) {
+                        throw new com.fran.pruebams.application.exceptions.FileAlreadyExistsException("File already exists");
+                    }
                     Files.copy(in, path);
                 } catch (IOException e) {
                     // Log this exception
